requirements.txt
Django>=3.2,<4.0
psycopg2>=2.8
djangorestframework>=3.14.0
django-composite-field>=1.1
django-cors-headers==3.13.0
django-filter>=21.1
django-rest-polymorphic
requests>=2.26
markdown>=3.4.1
drawSvg==1.9.0
django-celery-results
celery>=5.3


